pyproject.toml
[tool.poetry]
name = "learning-material-django"
version = "0.1.0"
description = ""
authors = ["Alicia Schonefeld <schonefeld.dev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
django-modeltranslation = "^0.16.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
